シェーダー読み取り機能 (HLSL)の実装 (テストはまだ行えていません.)

+++ PPP_GameEngine/GraphicsCore/RHI/InterfaceCore/Resource/Source/GPUBuffer.cpp
@@ -33,6 +33,6 @@ GPUBuffer::GPUBuffer(const std::shared_ptr<RHIDevice>& device, const core::GPUBu
 	/*-------------------------------------------------------------------
 	-          Set Stride and Element Count
 	---------------------------------------------------------------------*/
-	_metaData.Stride   = isConstantBuffer ? CalcConstantBufferByteSize(_metaData.Stride) : _metaData.Stride;
+	_metaData.Stride   = isConstantBuffer ? static_cast<size_t>(CalcConstantBufferByteSize(_metaData.Stride)) : _metaData.Stride;
 	_metaData.ByteSize = _metaData.Stride * _metaData.Count;
 }
